forgot to update some test files
